Made compatible with 6.4 (a bug arised)
With 6.4 a bug arised: when run in release mode and a line plot is
plotted, the line color is white instead of red, and stays white even if
a bar chart is then plotted.
If one plots bars as first plot, however, the bar have black contour as
expected. If the line color is inspected with qDegug(), it is, however,
red as expected.
In debug mode everyting is as expected, so finding this bug is very
difficult.
With Qt 5.15 this problem does not occur.